requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pytest==7.3.1
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.6.1
urllib3==2.0.2
